refactor: type 디렉터리로 정리
