+++ src/types/Book.ts
@@ -1,0 +1,15 @@
+export type Book = {
+  bookId: number;
+  bookInfoId: number;
+  status: number;
+  title: string;
+  author: string;
+  category: string;
+  categoryId: number;
+  isbn?: string;
+  publisher: string;
+  publishedAt?: string;
+  callSign: string;
+  image?: string;
+  isLendable: number;
+};
