+++ src/types/History.ts
@@ -1,0 +1,16 @@
+export type History = {
+  id: number;
+  lendingCondition: string;
+  login: string;
+  returningCondition: string;
+  penaltyDays: number;
+  title: string;
+  bookInfoId: number;
+  createdAt: string;
+  returnedAt: string;
+  dueDate: string;
+  lendingLibrarianNickName: string;
+  returningLibrarianNickname: string;
+  callSign: string;
+  image?: string;
+};
