+++ src/types/Lending.ts
@@ -1,0 +1,11 @@
+export type Lending = {
+  id: number;
+  title: string;
+  callSign: string;
+  dueDate: string;
+  createdAt: string;
+  lendingCondition: string;
+  login?: string;
+  penaltyDays: number;
+  image?: string;
+};
