+++ src/types/Reservation.ts
@@ -1,0 +1,13 @@
+export type Reservation = {
+  bookId: number;
+  callSign: string;
+  createdAt: string;
+  endAt: string;
+  image?: string;
+  login?: string;
+  penaltyDays: number;
+  reservationsId: number;
+  status: number;
+  title: string;
+  userId: number;
+};
