+++ src/types/Review.ts
@@ -1,0 +1,10 @@
+export type Review = {
+  reviewsId: number;
+  reviewerId: number;
+  bookInfoId: number;
+  content: string;
+  createdAt: string;
+  title: string;
+  nickname: string;
+  intraId: string;
+};
