+++ src/types/User.ts
@@ -1,0 +1,15 @@
+import { Lending } from "./Lending";
+import { Reservation } from "./Reservation";
+
+export type User = {
+  id: number;
+  email: string;
+  nickname: string;
+  intraId: number;
+  slack: string;
+  penaltyEndDate: string;
+  overDueDay: number;
+  role: number;
+  reservations: Reservation[];
+  lendings: Lending[];
+};
