+++ src/types/index.ts
@@ -1,0 +1,4 @@
+export * from "./User";
+export * from "./Lending";
+export * from "./Reservation";
+export * from "./Book";
